+++ sdk/core/src/shared/getDidData.ts
@@ -8,9 +8,14 @@ export const getDidDataFromKeysService = (keysService: KeysService) => {
   return { did, didDocumentKeyId }
 }
 
-export const withDidData = (userData: { encryptedSeed: string; password: string; didDocument?: any }) => {
-  const { encryptedSeed, password, didDocument } = userData
-  const keysService = new KeysService(encryptedSeed, password)
+export const withDidData = (userData: {
+  encryptedSeed: string
+  password: string
+  didDocument?: any
+  accountNumber?: number
+}) => {
+  const { encryptedSeed, password, didDocument, accountNumber } = userData
+  const keysService = new KeysService(encryptedSeed, password, accountNumber)
   const { did, didDocumentKeyId } = getDidDataFromKeysService(keysService)
-  return { did, didDocumentKeyId, encryptedSeed, password, didDocument }
+  return { did, didDocumentKeyId, encryptedSeed, password, didDocument, accountNumber }
 }